# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List the signed-in user's notifications, most recent first
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /resources:
    get:
      operationId: listResources
      tags: [resources]
      summary: Browse the resource library
      parameters:
        - $ref: "#/components/parameters/Search"
        - name: topic
          in: query
          schema:
            $ref: "#/components/schemas/ResourceTopic"
        - name: type
          in: query
          schema:
            $ref: "#/components/schemas/ResourceType"
      responses:
        "200":
          description: Resources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Resource"
  /resources/{id}/view:
    post:
      operationId: recordResourceView
      tags: [resources]
      summary: Record that a resource was opened (feeds admin "resource usage" analytics)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Recorded
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    Search:
      name: search
      in: query
      schema:
        type: string
...
  responses:
...
    Unauthorized:
      description: Not signed in
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    NotificationType:
      type: string
      enum: [reply, resource, maintenance]
    ResourceType:
      type: string
      enum: [article, book, video, podcast, verse, exercise, prompt, tip]
    ResourceTopic:
      type: string
      enum:
        - anxiety
        - grief
        - purpose
        - relationships
        - depression
        - stress
        - faith
        - healing
        - self-worth
...
    Notification:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          $ref: "#/components/schemas/NotificationType"
        title:
          type: string
        body:
          type: string
        conversationId:
          type: string
          format: uuid
          nullable: true
        read:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, type, title, body, conversationId, read, createdAt]
    Resource:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        type:
          $ref: "#/components/schemas/ResourceType"
        topic:
          $ref: "#/components/schemas/ResourceTopic"
        url:
          type: string
          nullable: true
        body:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, title, description, type, topic, url, body, createdAt]
...